Added game to readme.md
Added Realm of the Mad God as game in flash. RotMG was created in a dev contest in 2010. 
Last year the game was sold to the small developer studio DECA games, who actively try to this game back to life.
The game as a decent userbase and it is DECAs plan to keep it alive at least for another 10 years. However, they wouldn't have the resources to convert or rewrite this game into another language like HTML5. (Neither the money, nor the employees)